+++ app/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the cfxnes application will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Unrecoverable state after ROM with unsupported mapper is loaded.
+
 ## [0.7.0] - 2019-04-04
 
 ### Fixed
@@ -137,6 +143,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Initial version.
 
+[Unreleased]: https://github.com/jpikl/cfxnes/compare/v0.7.0...HEAD
 [0.7.0]: https://github.com/jpikl/cfxnes/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/jpikl/cfxnes/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/jpikl/cfxnes/compare/v0.4.0...v0.5.0